requirements.txt
torch
torchvision
scikit-learn
pillow
